Fully implemented second level grammar.
Removed tons of useless comments from skeleton file.

`InitAST` now is parent of both `AssignmentAST` and `VarBindingAST`. This is necessary to manage codeflow in for loop initialization, in case a variable is being declared or simply reassigned.

Implemented and tested logic for `IF` and `FOR` statements.

Might still improve a bit codegen methods of both classes if i can.